refactor: update drizzle config and move into new dir
