+++ drizzle.config.ts
@@ -2,7 +2,7 @@ import { defineConfig } from "drizzle-kit";
 import env from "./src/lib/envVariables";
 
 export default defineConfig({
-  schema: "./src/db/schema.ts",
+  schema: "./src/db/schema",
   out: "./supabase/migrations", // Storing migrations in supabase folder is a good practice
   dialect: "postgresql",
   dbCredentials: {
