+++ src/db/drizzle.ts
@@ -1,6 +1,6 @@
 import { drizzle } from "drizzle-orm/postgres-js";
 import postgres from "postgres";
-import * as schema from "./schema";
+import * as schema from "./schema/schema";
 import env from "@/lib/envVariables";
 
 // Disable prefetch for serverless environments usually, but fine for standalone
