+++ src/lib/context/VillageCodeContext.tsx
@@ -7,7 +7,7 @@ import {
 } from "react";
 import { trpc } from "@/utils/trpc";
 import { useSaveOnWrite } from "@/hooks/useSaveOnWrite";
-import type { VillageCode } from "@/db/schema";
+import type { VillageCode } from "@/db/schema/schema";
 
 interface VillageCodeContextValue {
   selectedVillageCodeId: number | null;
